Refactor ReadOnlyMiddleware as Graphene middleware; drop checks for Django views
