+++ saleor/core/middleware.py
@@ -1,22 +1,16 @@
 import logging
-import random
-import re
 
 from django.conf import settings
 from django.contrib.sites.models import Site
 from django.core.exceptions import MiddlewareNotUsed
-from django.http import JsonResponse
-from django.template.response import TemplateResponse
 from django.utils import timezone
 from django.utils.functional import SimpleLazyObject
 from django.utils.translation import get_language
 from django_countries.fields import Country
 
 from ..discount.utils import fetch_discounts
-from ..graphql.views import API_PATH, GraphQLView
 from ..plugins.manager import get_plugins_manager
 from . import analytics
-from .exceptions import ReadOnlyException
 from .utils import get_client_ip, get_country_by_ip, get_currency_for_country
 
 logger = logging.getLogger(__name__)
@@ -122,101 +116,3 @@ def plugins(get_response):
         return get_response(request)
 
     return _plugins_middleware
-
-
-class ReadOnlyMiddleware:
-    ALLOWED_MUTATIONS = [
-        "checkoutAddPromoCode",
-        "checkoutBillingAddressUpdate",
-        "checkoutComplete",
-        "checkoutCreate",
-        "checkoutCustomerAttach",
-        "checkoutCustomerDetach",
-        "checkoutEmailUpdate",
-        "checkoutLineDelete",
-        "checkoutLinesAdd",
-        "checkoutLinesUpdate",
-        "checkoutRemovePromoCode",
-        "checkoutPaymentCreate",
-        "checkoutShippingAddressUpdate",
-        "checkoutShippingMethodUpdate",
-        "checkoutUpdateMetadata",
-        "checkoutClearMetadata",
-        "checkoutUpdatePrivateMetadata",
-        "checkoutClearPrivateMetadata",
-        "tokenCreate",
-        "tokenVerify",
-    ]
-    BLOCKED_URL_PATTERNS = [
-        re.compile(r"^/dashboard"),
-        re.compile(r"^/([\w-]+/)?account/$"),
-        re.compile(r"^/([\w-]+/)?account/signup/$"),
-    ]
-
-    def __init__(self, get_response):
-        self.get_response = get_response
-
-    def __call__(self, request):
-        return self.get_response(request)
-
-    def process_view(self, request, *_args, **_kwargs):
-        if request.path == API_PATH:
-            if not self._is_graphql_request_blocked(request):
-                return None
-            error = GraphQLView.format_error(
-                ReadOnlyException("Be aware admin pirate! API runs in read-only mode!")
-            )
-            data = {"errors": [error], "data": None}
-            response = JsonResponse(data=data, safe=False)
-            response["Access-Control-Allow-Origin"] = settings.ALLOWED_GRAPHQL_ORIGINS
-            response["Access-Control-Allow-Methods"] = "POST, OPTIONS"
-            response[
-                "Access-Control-Allow-Headers"
-            ] = "Origin, Content-Type, Accept, Authorization"
-            return response
-
-        if self._is_django_request_blocked(request):
-            image = random.randrange(1, 6)
-            domain = Site.objects.get_current().domain
-            url = f"http://{domain}"
-            ctx = {
-                "image_path": "read_only/images/pirate-%s.svg" % image,
-                "image_class": "img%s" % image,
-                "back_url": request.headers.get("referer", url),
-            }
-
-            return TemplateResponse(request, "read_only/read_only_splash.html", ctx)
-
-    def _is_url_blocked(self, url):
-        for pattern in self.BLOCKED_URL_PATTERNS:
-            yield pattern.match(url)
-
-    def _is_django_request_blocked(self, request):
-        is_post = request.method == "POST"
-        request_url = request.path_info
-        return is_post and any(self._is_url_blocked(request_url))
-
-    def _is_graphql_request_blocked(self, request):
-
-        body = GraphQLView.parse_body(request)
-        if not isinstance(body, list):
-            body = [body]
-        for data in body:
-            query, _, _ = GraphQLView.get_graphql_params(request, data)
-            document, _ = GraphQLView().parse_query(query)
-            if not document:
-                return False
-
-            definitions = document.document_ast.definitions
-
-            for definition in definitions:
-                operation = getattr(definition, "operation", None)
-                if not operation or operation != "mutation":
-                    continue
-
-                for selection in definition.selection_set.selections:
-                    selection_name = str(selection.name.value)
-                    blocked = selection_name not in self.ALLOWED_MUTATIONS
-                    if blocked:
-                        return True
-        return False
