+++ saleor/graphql/middleware.py
@@ -9,6 +9,7 @@ from graphql import ResolveInfo
 from graphql_jwt.middleware import JSONWebTokenMiddleware
 
 from ..app.models import App
+from ..core.exceptions import ReadOnlyException
 from ..core.tracing import should_trace
 from .views import API_PATH, GraphQLView
 
@@ -56,6 +57,42 @@ def app_middleware(next, root, info, **kwargs):
                 if auth_prefix.lower() == prefix:
                     request.app = SimpleLazyObject(lambda: get_app(auth_token))
     return next(root, info, **kwargs)
+
+
+class ReadOnlyMiddleware:
+    ALLOWED_MUTATIONS = [
+        "checkoutAddPromoCode",
+        "checkoutBillingAddressUpdate",
+        "checkoutComplete",
+        "checkoutCreate",
+        "checkoutCustomerAttach",
+        "checkoutCustomerDetach",
+        "checkoutEmailUpdate",
+        "checkoutLineDelete",
+        "checkoutLinesAdd",
+        "checkoutLinesUpdate",
+        "checkoutRemovePromoCode",
+        "checkoutPaymentCreate",
+        "checkoutShippingAddressUpdate",
+        "checkoutShippingMethodUpdate",
+        "tokenCreate",
+        "tokenVerify",
+    ]
+
+    @staticmethod
+    def resolve(next_, root, info, **kwargs):
+        operation = info.operation.operation
+        if operation != "mutation":
+            return next_(root, info, **kwargs)
+
+        for selection in info.operation.selection_set.selections:
+            selection_name = str(selection.name.value)
+            blocked = selection_name not in ReadOnlyMiddleware.ALLOWED_MUTATIONS
+            if blocked:
+                raise ReadOnlyException(
+                    "Be aware admin pirate! API runs in read-only mode!"
+                )
+        return next_(root, info, **kwargs)
 
 
 def process_view(self, request, view_func, *args):
